Small connections.js refactor
