+++ tgui/packages/tgui/interfaces/IntegratedCircuit/constants.js
@@ -1,6 +1,5 @@
 export const NULL_REF = '[0x0]';
 export const ABSOLUTE_Y_OFFSET = -32;
-export const SVG_CURVE_INTENSITY = 64;
 
 export const MOUSE_BUTTON_LEFT = 0;
 export const OPTION_DROPDOWN_LARGE_CHAR_AMOUNT = 12;
